+++ packages/expo-router/src/react-navigation/routers/DrawerRouter.tsx
@@ -31,7 +31,7 @@ export type DrawerNavigationState<ParamList extends ParamListBase> = Omit<
   /**
    * Type of the router, in this case, it's drawer.
    */
-  type: 'drawer';
+  type?: 'drawer';
   /**
    * List of previously visited route keys and drawer open status.
    */
@@ -155,29 +155,13 @@ export function DrawerRouter({
 
     type: 'drawer',
 
-    getInitialState({ routeNames, routeParamList, routeGetIdList }) {
-      const state = router.getInitialState({
-        routeNames,
-        routeParamList,
-        routeGetIdList,
-      });
-
-      return {
-        ...state,
-        stale: false,
-        type: 'drawer',
-        key: `drawer-${nanoid()}`,
-      };
-    },
-
-    getRehydratedState(partialState, { routeNames, routeParamList, routeGetIdList }) {
+    getRehydratedState(partialState, { routeNames, routeGetIdList }) {
       if (partialState.stale === false) {
         return partialState;
       }
 
       let state = router.getRehydratedState(partialState, {
         routeNames,
-        routeParamList,
         routeGetIdList,
       });